[refactor] Lottery.tsx 파일 분리
